[FLINK-14202][table][python] Optimize the execution plan for Python Calc when there is a condition
Optimizes the Python UDF execution plan when there is a condition which contains
Python UDFs by calculating the condition firstly and then check whether to call
the other UDFs. This can eliminate the Python UDF execution for the input rows
which could be filtered by the condition.

Brief changes:
- Refactor PythonScalarFunctionSplitRule by splitting it into multi rules to
  make it more understandable.
- Handle cases where condition is not null in the SplitRule.

This closes #9907.